feat: add dailySummary method

+++ src/main/java/com/danieloliveira/order_management/dto/response/ResumoDoDiaResponseDTO.java
@@ -12,7 +12,7 @@ public class ResumoDoDiaResponseDTO {
 
     private LocalDate data;
 
-    private Integer quantidadePedidos;
+    private Long quantidadePedidos;
 
     private BigDecimal valorTotalEsperado;
 }
